+++ cluster/service/config.go
@@ -135,7 +135,7 @@ type Config struct {
 
 	// grpc service endpoint
 	SvrHost   string `toml:",omitempty"`
-	SvrPort   uint64 `toml:",omitempty"`
+	SvrPort   int    `toml:",omitempty"`
 	SvrModule string `toml:",omitempty"`
 
 	// Logger is a custom logger to use with the p2p.Server.
